Convert ike-test-utils from filegroup to a library
After jarjaring ike_test, CertUtils (in test code) is referencing
IkeCertUtils (in production code) by the new package name. The result
is any library that includes the ike-test-utils filegroup will fail
to resolve IkeCertUtils.

This CL converts ike-test-utils to a library that depends on the
jarjared IkeCertUtils file to solve this problem

Bug: 177266501
Test: CtsIkeTestCases verified with aosp/1573066(Allow IKE in BCP)
Change-Id: I0c6e3e764c735d44a5ea3b37c227b82f46c3921e