Re-init Locks and MemMap for each gtest
CommonRuntimeTest gtests will clear the MemMap statics when it deletes
the runtime. This can cause other tests to fail if you run the gtests
with the `--no_isolate` flag. By reinitializing these systems in
CommonArtTestImpl::Setup in addition to the original call in
gtest_main.cc this issue is avoided.

Test: $ANDROID_HOST_OUT/nativetest64/art_compiler_tests/art_compiler_tests --no_isolate

Change-Id: Ifb258e3ca6bf63ca0f821ed598143b3cc41d3e8b