Fix integer overflow in FindEntry.
Seems harmless because ReadAtOffset is careful, but let's not crash on
invalid input.

I've also renamed nameLen to name_length to match local style, and where
possible use .size() rather than the variable that was used to size the
vector.

Add a test that's basically the fuzzer output. This is the first bug
caught by that new fuzzer, which I wrote for last year's fuzzathon.
If/when we get more, I'll factor out the "reproduce an iteration fuzzer
crash" code, but as long as there's only one, it's just inline in the
test.

Bug: http://b/174945959
Test: new test
Change-Id: I2d63bc1c59218b534524efd642eeeb43e4cd1e7d