gsid: Make createPartition() don't preempttively clear installer_
As enableGsi() uses the nullness of installer_ to decide if an on-going
installation is happening, we shouldn't clear installer_ in
createPartition() no matter what.
Even when the installation failed to start, we should still leave
installer_ pointing to an useless installation session, so that the
caller can have a chance to handle the error state.
I think we probably need a dedicated variable to indicate the
installation state, so we can tell these subtle differences easier.

Bug: 165925766
Bug: 179980369
Test: Verify that DSU installation completes successfully even when
  scratch failed to be installed.
Change-Id: I7b465ac8a83859db45fc9c1a64eabfb76d780aee