Split C++ generated code into "enums" and "parser" parts
The "enums" part contains only the enum definitions from the
xsd file. This part is lightweight and doesn't depend on libxml2.
The "parser" complements the "enums" part and depends on it.

By default both parts are generated. To generate the "enums"
part only, enable "enums_only" property in the xsdc module.
To generate the "parser" part only, enable "parser_only".
These properties are mutually exclusive.

New unit tests added to verify that the decomposition works.

Bug: 180449205
Test: atest xsdc-cpp-tests xsdc-cpp-tests-split xsdc-cpp-tests-enums
Change-Id: I89833f23cdd2ed087b3258a4bc5deb68f95b47bf