Autoupdate of 2 submodules
*packages/modules/Virtualization updated to d8cf762504ce9f1289f88f253d40331f9ab15da3
Add SELinux stuffs to boot

Contexts files, plat_sepolicy.cil, and 10000.0.cil are needed to boot.
This adds those files to microdroid. But cil files are temporary and
only for testing. We'll need to migrate real cil files to Android.bp.

Bug: 178993690
Test: boot microdroid
Change-Id: I684b67bbb6906ba42a710156519723eff23938d2

*system/sepolicy updated to e35b49bd1680d56d771389d5fc7f915a2b553e89
Add cil files to Android.bp for microdroid

Contexts files, plat_sepolicy.cil, and 10000.0.cil are needed to boot.
This adds cil files to microdroid. But cil files are temporary and only
for testing. We'll need to migrate real cil files to Android.bp.

Bug: 178993690
Test: boot microdroid
Change-Id: I711b1db39c11d88bc1f9defeff5799e6f24756ab