Fix the mac build
mkenvimage_host was enabled only for linux and this caused a build
failure on mac. Although we don't expect we will build uboot_env on a
mac, the build system comlains when it finds a dependency to a disabled
tool from a genrule module. Fulfill the requirement by providing a fake
binary of mkenvimage_host for mac.

Bug: 180973515
Test: `m nothing` on mac
Change-Id: Iff0add1c5e45ace48ced3c90bfde29a5d376d7c5