Move toolchain def to WORKSPACE
This fixes an issue with mixed builds where the toolchain location was
invalidly specified; if specified in the bazelrc, then the toolchain was
defined under the *current* repository, but would have accurately been
resolvable under @sourceroot.

A simple bazelrc fix (to sourrceroot location) would not have sufficed,
as that would break bazel builds outside of mixed-build context.

Test: Building crtbegin_so1 using mixed builds and just using bazel,
after incorporating aosp/1599675

Change-Id: I5f4815770d9f84e922688b757034d3cfb4852568