Warn in headers that this is a fork.
There have been a number of times where someone has tried to use this
with AIDL code (or vice-versa). Since libhwbiner is the fork, and it has
far fewer headers, and since it will eventually be deprecated/deleted,
I've chosen to add the warning here (adding it to libbinder would also
mean removing it there).

For people confronted with two copies of this classes, beyond their
different include paths, this should help explain the difference.

Fixes: 180746712
Test: N/A
Change-Id: I91fec179a54430da03108c8cd4d25aa18e7ec361