hidl-gen: don't format if it'll destroy comments
hidl-gen has a limited capacity to handle comments, since it always
tries to put them in the right place. Ideally, it could still handle
these (large reworking of the AST/formatter/etc, it seems) or it would
have had these restrictions on comment placement from the beginning
(it's an interface file! how about we always know what thing a human
wants to say something about, so that what the human wants to say about
can be transformed when the source is transformed).

Because of this, frequently when running the formatter, comments can get
destroyed, and it can be easy to miss these. Now, a user must manually
remove/fixup these comments before we destroy any information.

Test: hidl-gen -Lformat android.hardware.neuralnetworks@1.2::IDevice
        (succeeds for one file)
Test: hidl-gen -Lformat android.hardware.neuralnetworks@1.2::types
        (and you get warnings from one file)
Test: hidl-gen -Lformat android.hardware.neuralnetworks@1.2
        (and you get warnings from the whole package)

Fixes: 180899879
Change-Id: Icce1262425fa70006fe5dea5ce0271ffb2b9492f