Add SMT flag and disable by default
Exposing SMT/HT to the guest can cause some scheduling decisions to be
made by the guest that can reduce net performance. This change allows
the CPUs to be exposed in a flat / non-SMT topology.

Modern crosvm defaults to enabling SMT, and has a -no-smt flag to turn
it off, but our existing QEMU code was exposing the CPUs as separate
sockets, which disables SMT. Align both backends to use cores/threads
instead of sockets, and disable SMT by default for crosvm. Then, allow
SMT to be enabled again for both crosvm and QEMU.

Note that although e.g. "cores=4,threads=2" gets passed to QEMU now, it
seems to ignore it and always presents the CPUs in a flat topology. This
may be a QEMU bug or a limitation of the host CPU I tried it on.

Bug: 181582092
Change-Id: Ic0888c84d101e31d33cd6ed5fc82b1440ab6d0e9