Fork rules_cc starlark rules (and deps)
This institutes a temporary fork of
https://github.com/bazelbuild/rules_cc and one of its dependencies
(https://github.com/bazelbuild/bazel-skylib). This fork contains a
minimal set of a features to facilitate prototyping of Bazel builds of
C++ library targets and the toolchains they depend on.

At a later time, we will use a more principlied solution to pull HEAD of
these two github repositories into AOSP.

Test: Manually verified use of experimental_shared_library.bzl in a
separate prototype change.
Test: Treehugger

Change-Id: I9b35c3b15b7cce408c8e7f87e76b8e509345771a