Autoupdate of 3 submodules
*hardware/interfaces updated to 1f376685524cc54c7c36ffc1f7bd8e6ce8ba1ba4
Merge "neuralnetworks: fix out ParcelFileDescriptor"
*packages/modules/NeuralNetworks updated to cf5dff728654c43611b347ad62763b5248ffb8b9
Merge "neuralnetworks: fix out ParcelFileDescriptor"
*system/tools/aidl updated to 020d8d12caa98282ab69b5dc5cd1b7be6d2f6af4
Reject "out ParcelFileDescriptor" parameters

Out(not inout) parameters should be known how to create it. But in Java,
ParcelFileDescriptor is not default constructible and there's no way to
construct it from nothing.

Instead of rejecting it only in Java, we reject it in every backend for
future compatibility.

Bug: 181194872
Test: aidl_unittests
Change-Id: I93637bb6f5bbbd9f4c3119c0baa8ba4042fc2305