Autoupdate of 2 submodules
*art updated to 21dafaf33ed6d7d6df1ee00371a3503802a060f5
Skip make targets that have missing dependencies when using ART
prebuilts.

This is necessary because ART sources can still be in the tree when
prebuilts are used, and e.g. for host we don't have a boot image in
them.

Test: m checkbuild
  with and without SOONG_CONFIG_art_module_source_build=false, check
  that it passes the "Test for dangling rules" stage
Bug: 172480615
Change-Id: I4f24a430d1b11202eb4dceeedc96a0906679830e

*build/make updated to 36c1435ec4ac40dc273c916028d7f818f71ae9f3
Merge "Include android_soong_config_vars.mk in the make (kati) step."