Add stubs for rules_cc defs.
This is a workaround to fix an issue introduced by the rules_cc fork,
a partial checkin of the repository. While building //bionic/... on
bp2build BUILD files, a transitive dependency to the missing defs.bzl
through rules_cc resulted in broken load statement.

See attached bug for more information. This CL adds the def.bzl file to
fix the bug temporarily while rules_cc is still a fork.

Fixes: 181706395
Test: tested both bp2build and mixed builds
Test: build/bazel/scripts/milestone-2/demo.sh {generate, sync} && bazel build //bionic/... && USE_BAZEL=1 USE_BAZEL_ANALYSIS=1 m libc

Change-Id: I27eb818a3a56b25003e1f3f058c8251d507c373b