Autoupdate of 2 submodules
*ndk updated to 178f38b99fffd0ddac81ec9ec97b482f09692dba
Merge changes from topic "ndk-libunwind-res-dir"

* changes:
  Switch to the arm32 libunwind.a from the resource dir
  Reset linker_ldflags in build-binary.mk

*toolchain/llvm-project updated to e995111c3b40453d6ec16b686d030d13cc1f788a
Switch to the arm32 libunwind.a from the resource dir

Remove the -ldl because the Clang driver automatically adds -ldl when
it's needed for unwinding (dynamic bin/libs), and omits it from static
executables where the unwinder only needs libc.a.

Define the libunwind.a prebuilt for all architectures, even though
it's currently only used for arm32. We're planning to use it on all
architectures soon.

Bug: http://b/153025717
Test: checkbuild.py && run_tests.py
Change-Id: I03e74402ba1c42357691493015e7e39830ca8497