perfetto: decouple statsd monitoring from incidentd and allow skipping incidentd
This CL does two things:
1. Adds an explicit field for enabling/disabling logging to statsd
instead or relying on the presence of incidentd/enabling extra
guardrails in the service. This reduces subtlety and overloading of the
incidentreportconfig in the config proto.
By default, if unset this field will be set based on the incidentreportconfig
for backward compatibility.
2. Allows disabling saving to incidentd when configs come statsd. This
is useful for configs which only set bugreport_score and we don't want
to actually collect them.

Change-Id: I435b3202a61f0df736d1299a8fbd7dcd7a484cff