Autoupdate of 3 submodules
*bionic updated to 03e961e392b5d0761eb604ef2d8e9e8387ff40c9
Merge "Teach debuggerd to pass the secondary ring buffer to __scudo_get_error_info()."
*external/scudo updated to 75ee3bc68036ee471f25955ae72a533540dcf679
scudo: Add support for tracking stack traces of secondary allocations.

There is no centralized store of information related to secondary
allocations. Moreover the allocations themselves become inaccessible
when the allocation is freed in order to implement UAF detection,
so we can't store information there to be used in case of UAF
anyway.

Therefore our storage location for tracking stack traces of secondary
allocations is a ring buffer. The ring buffer is copied to the process
creating the crash dump when a fault occurs.

The ring buffer is also used to store stack traces for primary
deallocations. Stack traces for primary allocations continue to be
stored inline.

In order to support the scenario where an access to the ring buffer
is interrupted by a concurrently occurring crash, the ring buffer is
accessed in a lock-free manner.

Differential Revision: https://reviews.llvm.org/D94212

GitOrigin-RevId: 1f55fa0b99e0755c96bfd5cd0cdbe9cbc26fe2db
Change-Id: I56d79938555281e871f3ceb7654742c9b0f263b1

*system/core updated to 90947d442cb1ae59474f674ba663d8ba2dfdf554
Merge "Teach debuggerd to pass the secondary ring buffer to __scudo_get_error_info()."