Autoupdate of 3 submodules
*external/sl4a updated to ca96feb025f4c06083dc69f797e94cc72e53d4f7
Remove usage of StringNetworkSpecifier

StringNetworkSpecifier is a hidden API class that is removed in R, and
replaced by EthernetNetworkSpecifier and TestNetworkSpecifier.

Remove its usage in ConnectivityManagerFacade. It was misused to hold
JSON data to be later parsed and converted to a wifi NetworkSpecifier.

The current NetworkRequest builder from JSON is only used in
connectivityRegisterNetworkCallback which is unused, in
connectivityRequestNetwork which is never used with a NetworkSpecifier,
and in connectivityRequestWifiAwareNetwork. Only keep NetworkSpecifier
support in connectivityRequestWifiAwareNetwork, as the JSON format only
allows specifiers for that kind of network (generic string specifiers
should not be used as they are considered as EthernetNetworkSpecifier,
and the JSON format does not define a NetworkSpecifier type).

Bug: 179329291
Test: m sl4a
Merged-In: Iaa7a5395cd6f2af7f788b25276d8ba6c4c0eb2a2

Change-Id: Iaa7a5395cd6f2af7f788b25276d8ba6c4c0eb2a2

*frameworks/base updated to c77293b7a92590e5583b9af0e4a6128d841f8c6f
Merge changes from topic "ethernet_specifier"

* changes:
  Fix common tests on Q and R
  Add Ethernet, TestNetworkSpecifier API

*frameworks/opt/net/ethernet updated to 244e39c0dec83714d065b69de38de96e0de1a950
Merge "Rename StringNetworkSpecifier to Ethernet"