Autoupdate of 2 submodules
*build/blueprint updated to ddb5ed7e1f4ad3c47e951a17fe24f98be0c949c7
Add ShellEscapeIncludingSpaces(string)

ShellEscape(string) doesn't escape a string with spaces like "arg1
arg2". However, when we want to escape a string to use it as an
argument, then strings with spaces should be escaped.

For example,

  "command " + ShellEscapeIncludingSpaces("a b")

becomes "command 'a b'" so that "command" will get "a b" as a single
argument.

Bug: 182092664
Test: Added tests to escape_test.go
Change-Id: I8f88c18bc4f9f7aacfe9e701b8f0876dd8b9a8c3

*build/soong updated to d0b3574072fec2e559e39040c086f4b2b23090d7
Merge "Fix bootimg.cmdline to use ShellEscapeIncludingSpaces"