Fix incompatability by swapping KE and Nonce in IKE INIT
The old IKE spec RFC 4306 requires the payload order in IKE INIT
to be SAi, KEi, Ni and allow responders to reject requests with
wrong order. Although starting from RFC 5996, the protocol removed
the allowance for rejecting messages in which the payloads were not
in the "right" order, there are few responder implementations
are still following the old spec when handling IKE INIT request
containing COOKIE payload. Thus IKE library should follow
the payload order to be compatible with older implementations.

Bug: 174997213
Test: FrameworksIkeTests (new tests added), CtsIkeTestCases
Change-Id: Id8a981f2d9284a41a8cd610abc1346f1fa26001e