Revert^2 "[metrics] Always enable background thread"
Having the thread enabled in some cases and not in others was causing
issues this tests. To avoid needing to handle multiple configurations,
this CL enables the background reporting thread always. In general,
metrics reporting will be enabled, so this better reflects the default
state of things.

This reverts commit 3f64940d7c6ce4d3f06ea5dbe195c244ca7a8fbc.
Reason for revert: Deflaking tests

The issue was that sometimes the metrics background thread would start
late and get captured by the thread life cycle callbacks instead of
the correct thread. The fix is to request a synchronous metrics report
before starting the test, which forces us to wait until the metrics
thread is up and running.

Test: ./test.py --run-test --host
Test: art_runtime_tests \
    --gtest_filter="ThreadLifecycleCallbackRuntimeCallbacksTest.*"
Bug: 170149255
Change-Id: I007437abd1a3404960f79cf1596c95b8da917286