VTS: test for kfence.sample_interval=500
kfence.sample_interval MUST be 500 for all 5.10 kernels and
is STRONGLY RECOMMENDED for 5.4 kernels.

Because the sample interval can be set via either
CONFIG_KFENCE_SAMPLE_INTERVAL or kfence.sample_interval, we check for
the module parameter value stored in /sys/module/kfence/parameters/sample_interval
The check is only performed for 5.10 kernels.

Bug: 181118352
Bug: 176230369
Test: atest KernelApiSysfsTest

Signed-off-by: Alexander Potapenko <glider@google.com>
Change-Id: Icac1f6d43f4891552842c265ec8ceb07f7bb26a4