Autoupdate of 2 submodules
*build/blueprint updated to b3b9cb6ccc67a88bb33d359fb6b76f55b5d54ba5
Remove the GeneratingPrimaryBuilder() method.

It's replaced with a flag in bootstrap.Args. It's a little bit ugly that
bootstrap.Main also gained a new Boolean flag in the process. The plan
is to delete that method in favor of directly calling RunBlueprint().

Test: "m nothing" (presubmits take a full workday these days)
Change-Id: Iae3421ae517a90570259b6278f18c3614bb1bb73

*build/soong updated to d7ce8402d8895b7ba9d13092d27037784a43cc0e
Remove GeneratingPrimaryBuilder().

It's replaced with a flag in bootstrap.Args .

Test: "m nothing" (presubmits take a full workday these days)
Change-Id: Ia0bfa2091013e389890e583c559006077806af99