Autoupdate of 2 submodules
*hardware/interfaces updated to 72ca1d2a401b1ead4268d52ae725800a757c4948
Merge "COSE unprotected parameters are a map not a bstr"
*packages/apps/RemoteProvisioner updated to 00599d0a8c42d79cccef89ef0d99c6132bb16d54
Fixing mac0 unprotected parameter encodings

This change fixes the structure of the COSE_mac0 structures created in
the test packages to match the spec. The unprotected parameter should
always be encoded as a map, not as a bstr if it's empty.

Test: atest RemoteProvisionerUnitTests
Change-Id: Ib29257faadf33d6c2e83e56d119d576cd7986ac7