Exempt VNDK apex from duplicate check
Development builds of aosp_$pixel contain a VNDK apex in /vendor,
because the VNDK ABI interface is not stablized.
If the system.img and vendor.img are built separately, for example
system is GSI, then there would be two VNDK apexes on the device.
One in /system and one in /vendor.
In this case only the one in /system would be activated and the
/vendor one would be ignored.
Exempt this case from the duplicate test.

Bug: 182418690
Test: Flash aosp_flame vendor.img and GSI system.img, then boot to home
  screen. (Development branch)
Change-Id: I3bf7a2dab410e3078151ad7da2a5b574c4be74d0