Remove kAccMiranda and kAccDefaultConflicting.
Use a combination of other existing flags to mark miranda
and default conflicting methods.

Also change the value of identical flags kAccCriticalNative
and kAccPreCompiled to 0x00100000 to have the values of
kAccFastNative and kAccCriticalNative in consecutive bits.

The method access flag 0x00200000 is now free for future
use. The method access flag 0x00100000 is also potentially
usable for some methods the same way that the old
kAccMiranda flag was used despite sharing the same value
with both kAccCriticalNative and kAccPreCompiled.

Test: m test-art-host-gtest
Test: testrunner.py --host --optimizing
Bug: 112676029
Change-Id: Ibfd20598f9b4af78e1e72f8331dc647d01dd8eb5