ScopedIcuLocale incorrectly parses the input as locale ID
All usages of ScopedIcuLocale are in LocaleNative, CaseMapperNative,
TimeZoneNamesNative, and they provided a BCP-47 language tag
instead of ICU locale ID. ScopedIcuLocale should parse the input
as language tag.

The tests for those 3 classes are added in the prior CLs.

The bug wasn't an issue because Locale::createFromName is lenient enought
to parse language tag, espesically accepting both - and _ as the
separator. However, it's not guaranteed in the future ICU version.

No known regression is caused by this bug.

Test: CtsLibcoreTestCases
Test: CtsIcuTestCases
Change-Id: Iea107619d638e06d58623b702793959df1210e68