Add unit test cases for multithreaded access to BufferAllocator
They will verify concurrent access of a BufferAllocator object.

The patch also contains makefile changes to generate different versions
of the test for VTS and presubmit since all of the tests can be run in
presubmit but only a subset of them can be used for compliance(since ION
did not have stable ABI).

Test: atest dmabufheap-unit-tests
Test: atest dmabufheap-vts-unit-tests
Bug: 154310076
Change-Id: I7f2ca84a0f867fcc005a5a2cf372ed2509676893