Fix Java ReverseParcelFileDescriptorArray tests.
This was failing before because PFDs are automatically closed when they
are returned from binder transactions.

This behavior is not ideal because it could close PFDs that we intend to
keep around. Changing this behavior could be problematic, and adding a
way to control whether PARCELABLE_WRITE_RETURN_VALUE is affected adds
too much complexity (especially since returning the same PFD multiple
times isn't expected).

If an extra dup like this became problematic, we could add
ParcelFileDescriptor.ignoreCloseOnWrite. Though, it would be really
really messy. For now, updating the test to explicitly show Java vs
regular behavior.

Fixes: 178863692
Test: ./run_integration_test.sh
Change-Id: If3db4718b118bbe2120e9515a1a500748cc2c3c3