Allow individual directories to be mounted in nsjail.
Until now the Keystone overlays mounted git projects individually and only
mounted the root for regular non-git-project directories. Thus if a directory
structure was:

   dir1/
      dir2/
      dir3/

and none of those were git projects, then only dir1 would be mounted.

This change allows a directory to be mounted by adding a .bindmount file,
for instance:

   dir1/
      dir2/
         .bindmount
      dir3/

would cause dir2 to be mounted individually, and also dir3, but not dir1.

This change enables preventing conflicts when a directory being mounted
would be a prefix to a different project also being mounted, for instance:

   dir1/
      .git
      ...
   overlays/
      dir1/
         dir2/
      dir3/
         .git

In the example above the dir1 project and the overlays/dir1 directory would
be mounted. To prevent conflicts, one could directly mount dir2 like this:

   dir1/
      .git
      ...
   overlays/
      dir1/
         dir2/
	    .bindmount
      dir3/
         .git

Bug: 178194854
Bug: 183439458
Test: ./test.sh
Change-Id: I31122765ac316aea74b2c9c6aadcc7820d52ed97