Convert aidl package to use test fixtures
Converts the aidl_test.go file to use the new test fixture mechanism.
Usually, the preparers would be added altogether in a single
GroupFixturePreparers() or maybe a few separate ones. However, in this
case they are appended to a list roughly in the same order as the
_aidlTest() function previously initialized the test context and
config. That should make it easier to correlate the previous code with
the new code.

Generally, the order in which they are added does not matter, as the
test infrastructure automatically makes sure that mutators,
pre-singletons and singletons are run in the correct order and the rest
of the registration code does not care. However, the setting of the
product variables to their default values must be added before the
customizers passed into _aidlTest() as the latter can override the
defaults.

This also removes the buildDir and the setUp/tearDown boilerplate as
that is no longer needed. The one test that used it has been switched
to use RelativeToTop() which will convert absolute paths in the
temporary test build directory into paths relative to
ANDROID_BUILD_TOP.

This change was made because upcoming changes to the Soong test
infrastructure would break this test otherwise. There is plenty of
additional cleanup that can be done, e.g. switching to use the new
android.Assert... methods but that is out of the scope of this change.

Bug: 181070625
Test: m nothing
Change-Id: I9e3d6a2573ed15bce36dbb0f2161de4da800453e