Autoupdate of 2 submodules
*frameworks/base updated to 2108a92452004f0115bca8cfdcba3435e9391452
Merge changes from topic "blocked-reasons-callback-tests"

* changes:
  Add unit tests for onBlockedStatusChanged(Network, int).
  Add onBlockedStatusChanged(Network, int) to NetworkCallback.
  Add MOBILE_DATA_PREFERRED_APPS setting
  Add PRIVATE_DNS_MODE setting setter
  Expose ConnectivitySettingsManager as module-lib API

*frameworks/libs/net updated to 5d4e7a5d364760be726762043dc6806502ae018f
Make TestableNetworkCallback understand blocked reasons.

This provides some support for tests that want to test the new
onBlockedStatusChanged(Network, int) callback. This cannot be
implemented entirely in TestableNetworkCallback, because in
Kotlin, "override" is a mandatory keyword, but NetworkCallback
(which TestableNetworkCallback inherits from) only has the new
method in S, not R.

The unit tests in the other CL in this topic provide an idea of
how to do this. There may be more elegant ways to do it though.

Bug: 165835257
Test: new ConnectivityServiceTest tests in this topic
Change-Id: Ibdf36a6bc916bf247688c19bfd81f6b62d59bf03