hidl2aidl: stop expanding types from another package
Added '-e' parameter to allow the old behavior.
Major and minor versions are ignored for this check.
If an interface or enum inherits from a type outside of its package, it
is no longer expanded in the new generated aidl files.
If a compound type includes another type from outside of its package, it
is no longer expanded in the new generated aidl files.

Test: hidl2aidl -o out -rhidl2aidl.test:system/tools/hidl/hidl2aidl/test hidl2aidl.test.extension@1.2
Test: With and without '-e' argument to verify expansion behavior
Bug: 183538171

Change-Id: I494ae28b4e8f8ecaacb94c41be8085baad42fa14