Autoupdate of 2 submodules
*build/soong updated to 58367574a593f76d6d32a0251679046eb28e4110
Merge "Add the transitive dependencies of boot_image to apex"
*packages/modules/RuntimeI18n updated to a870732da5e97ae542388a9e48e6cc3127cbc894
Add i18n-bootclasspath-fragment to apex

Adds a bootclasspath_fragment that wraps the core-i18n library to the
com.android.i18n apex.

Removes the core-icu4j library from the apex as that will be
transitively included via the bootclasspath fragment.

Bug: 177892522
Test: m com.android.i18n
      deapexer extract out/target/product/generic_arm64/system/apex/com.android.i18n.apex i18n
      Check i18n directory to make sure that core-icu4j is in there and there are no .art files either.
Change-Id: I893d75c53c2d3c0456d7ab265a0dba98cbf6160c