Improve libcppbor parser fuzzing
The parser should be fuzzed with arbitrary inputs including valid and
invaild examples. The corpus contains valid examples of encoded CBOR so
only use that for fuzzing the parser and directly pass the fuzz input
data to the parser.

libcppbor has a number of slightly different APIs for the parser so use
a number of them to parse the same input to get wider coverage.

Test: SANITIZE_HOST=address SANITIZE_TARGET=hwaddress make cppbor_parser_fuzzer
      && cppbor_parser_fuzzer corpus/
Change-Id: I9fc9b979fa21295ead9f6ae363e9f380009ac819