Stop leaking FD in AidlMessageQueue on error
If a newly created AidlMessageQueue was initialized with
numElementsInQueue smaller than SIZE_MAX / sizeof(T), but larger than
std::numeric_limits<int32_t>::max(), then the AidlMQDescriptorShim
constructor would return before taking ownership of the native_handle_t
being passed in.
The AidlMQDescriptorShim deconstructer would then not be able to
close/delete it like it is supposed to. This behavior leaks that FD.

This CL makes sure the AidlMQDescriptorShim takes ownership of the
native_handle_t before doing anything else.

This was found while working on a fuzzer for libfmq.

Test: atest fmq_unit_tests
Bug: 165621851
Change-Id: I92f3847bf34f76b8ce19cfccac8bbda21edb3922