Autoupdate of 2 submodules
*device/google/cuttlefish updated to 352900c39b10359b7d967d64b65d39a44f9ebc0c
Merge "Configure NNAPI updatable shell service"
*packages/modules/NeuralNetworks updated to 762161e47033211946653e5aab189a27a6b24449
Create first verison of the NNAPI updatable shell service

This first version hardcodes some values such as the service metadata
(number of cache files, capabilities, etc.) as well as the full support
library binary path.

This CL includes a fallback "InvalidDevice" for when the support library
is not able to be located, opened, or used. To test this functionality,
modify the name of the prebuilt path to an invalid path or change the
function name "ANeuralNetworks_getSLDriverImpl" to something else.

Bug: 170696191
Test: VtsHalNeuralnetworksTargetTest
Change-Id: I882e520b33d527ab1a1e212e2c24b08df3723994