Autoupdate of 2 submodules
*external/crosvm updated to 38a78326ce3cfe429a8344297abfb162d47e6204
crosvm prefers rlib even on device

crosvm is installed to the virt APEX, and there are only two rust
binaries (crosvm and virtmanager) in the APEX. Static linking is more
beneficial than dynamic linking when the number of binaries are such
low.

Bug: 181751814
Test: TARGET_BUILD_APPS=com.android.virt m
Size is reduced from 180MB to 145MB

Change-Id: I0f71eab76dee40a6e7be18eaa979895abfd087e6

*packages/modules/Virtualization updated to 26bdf449272ebb80dd3220300b5e203eccd0cbf5
virtmanager prefers rlib even on device

virtmanager is installed to the virt APEX, and there are only two rust
binaries (crosvm and virtmanager) in the APEX. Static linking is more
beneficial than dynamic linking when the number of binaries are such
low.

Bug: 181751814
Test: TARGET_BUILD_APPS=com.android.virt m
Size is reduced from 180MB to 145MB

Change-Id: I50d77adc1880e36a712a7f1a427d6d3ab0fa0a0b